Revert "SCRD-6921 - update SUSE theme from UX team feedback"
The original changes were based on the eos design system,
which is intended for applications and does not match
what the Documentation team is currently producing.

An updated set of changes will likely be needed to get
a closer match to the docbook xml theme that is used
for most SUSE docs, but if nothing else the pre-change
version of this theme is a much closer match.

updated fontawesome class names from fa to fas for FA5 compatability

This reverts commit 030fa0529d37ebca228094b76021ade6318038b3.